+++ TgAdsStatistics/Models/Post.cs
@@ -9,11 +9,11 @@ namespace TgAdsStatistics.Models
     {
         public int Id { get; set; }
         public string Date { get; set; }
-        public float Views { get; set; }
-        public float Subscribers { get; set; }
-        public float Cost { get; set; }
-        public float Convercy { get; set; }
-        public float SingleSubscriberCost { get; set; }
+        public int Views { get; set; }
+        public int Subscribers { get; set; }
+        public int Cost { get; set; }
+        public int Convercy { get; set; }
+        public int SingleSubscriberCost { get; set; }
         public int? ChannelId { get; set; }
         public Channel Channel { get; set; }
     }
